+++ src/common/serial.c
@@ -431,7 +431,7 @@ int serial_write(int fnbr, const char *buf, size_t sz) {
     assert(file_table[fnbr].type == fet_serial);
     errno = 0;
     ssize_t count = write(file_table[fnbr].serial_fd, buf, sz);
-    if (count == sz) {
+    if (count == (ssize_t) sz) {
         return sz;
     } else if (count == -1) {
         error_throw(errno);
